Refactor and clean up codebase
- Removed commented-out dependency for Spring Boot Quartz from pom.xml.
- Cleaned up unused imports in Blog.java, JobCrawlRepository.java, and CrawlServiceImpl.java.
- Added missing @Override annotation in WebMvcConfig.java.
- Updated CrawlServiceImpl.java to use constants for class names instead of hardcoded strings.
- Modified FirebaseConfigController.java and FirebaseHistoryController.java to improve variable declarations.
- Enhanced FirebaseStorageService.java and its implementation to support GCS uploads.
- Updated TranscriptController.java to handle RevAI API client initialization.
- Refactored GoogleTranscriptService.java to use CompletableFuture for asynchronous processing.
- Improved RevAITranscriptService.java to include API client as a parameter and refactor job status handling.
- Added constants in Constant.java for better maintainability.
- Fixed potential issues in FileSizeUtils.java by ensuring proper type handling.
- Updated application.yml to configure Quartz job store type and improve Spring Boot Admin URLs.

+++ src/main/java/com/app/toeic/util/FileSizeUtils.java
@@ -9,6 +9,6 @@ import java.math.RoundingMode;
 @UtilityClass
 public class FileSizeUtils {
     public BigDecimal convertByteToMB(BigInteger byteSize) {
-        return new BigDecimal(byteSize).divide(BigDecimal.valueOf(1024 * 1024), 2, RoundingMode.HALF_UP);
+        return new BigDecimal(byteSize).divide(BigDecimal.valueOf((long) 1024 * 1024), 2, RoundingMode.HALF_UP);
     }
 }
